+++ tests/test_admin.py
@@ -98,14 +98,6 @@ class AliasContentManagerTestCase(CMSTestCase):
             expected_en_content.name,
             response_content_decoded,
         )
-        self.assertIn(
-            expected_en_content.get_absolute_url(),
-            response_content_decoded,
-        )
-        self.assertInHTML(
-            'Actions',
-            response_content_decoded,
-        )
 
     @skipUnless(is_versioning_enabled(), 'Test only relevant for versioning')
     def test_alias_content_manager_rendering_with_versioning_actions(self):
@@ -183,3 +175,4 @@ class AliasContentManagerTestCase(CMSTestCase):
             expected_en_content.get_absolute_url(),
             response_content_decoded,
         )
+